requirements.txt
django>=1.7
django-appconf
django-compressor
django-extras
djangorestframework

